% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_nuts_osm_sf}
\alias{de_nuts_osm_sf}
\title{German NUTS regions}
\format{
An \link[sf:sf]{sf} object with 444 features and 4 fields:
\describe{
\itemize{
\item name: Name of the region
\item nuts_1, nuts_2, nuts_3: Official region code at NUTS-1, 2, and 3
}
}
}
\source{
\url{https://www.openstreetmap.org}, \url{https://overpass-api.de/api/interpreter}, \url{https://nominatim.openstreetmap.org/details}
}
\usage{
de_nuts_osm_sf
}
\description{
A dataset containing the names, NUTS codes, and geographic geometries of
German administrative regions
}
\keyword{datasets}
