% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{nuts_geocode}
\alias{nuts_geocode}
\title{Get NUTS codes for location strings}
\usage{
nuts_geocode(
  locations,
  nominatim_api = "https://nominatim.openstreetmap.org",
  return_geometry = FALSE
)
}
\arguments{
\item{locations}{Character vector of locations to geocode}

\item{nominatim_api}{URL of the Nominatim API}

\item{return_geometry}{If TRUE, returns the sf geometry of the geocoded NUTS
regions}
}
\value{
A data.frame with 5 columns:
\itemize{
\item location: The input in the \code{location} argument
\item name: Name of the geocoded NUTS region
\item nuts_1, nuts_2, nuts_3: Official region code at NUTS-1, 2, and 3
}
}
\description{
Get NUTS codes for location strings
}
\examples{
nuts_geocode(c("munich", "hamburg"))
}
